Add python portion of the project
Notes:

Added recognize.py. This initial python file has two functions:

- Analyze Text. This function will analyze and return PII from the
text by using the hbs prompts defined in the repo.

- Redact. This function will use the analyze function and attempt a
redaction based on that.

Added demo.py. This initial file provides a test UI driver for PII
Shield.

- User can put in OpenAi key. After entering text, redaction will
be done on the data.

Todo:

- Create a test driver. There are two approaches
 - Use all prompts against all test
 - Use relevant prompt against relevant test

Approach 1 is costlier, but more likely to mirror real world usage.
This would require implementing a harness over recognize.py to
run on all the tests.

Approach 2 is inexpensive, but may represent a higher accuracy
than we see in production in the lib.

May try approach 1 first in order to test the limited evals we
have.